refactor: moved constants to its own folder
