+++ src/estimate_distance.py
@@ -1,4 +1,4 @@
-from constants import CVZONE_DETECTOR_MAX_ONE, EYE_DISTANCE_INDEX, FOCAL_LENGTH, INTEROCULAR_DISTANCE
+from constants.model_constants import CVZONE_DETECTOR_MAX_ONE, EYE_DISTANCE_INDEX, FOCAL_LENGTH, INTEROCULAR_DISTANCE
 
 
 def get_z_estimation(face):
