+++ src/main.py
@@ -13,7 +13,7 @@ from estimate_distance import get_z_estimation
 from models.code.cvzone import detect_face_cvzone
 
 # Importing the instance of detector
-from constants import CVZONE_DETECTOR_MAX_ONE
+from constants.model_constants import CVZONE_DETECTOR_MAX_ONE
 
 # Setup for the information for the UDP server. 
 sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
